# lib/api-zod/src/generated/types/createNewsArticle.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateNewsArticle {
  /** Display date, e.g. 'May 27, 2026' */
  date: string;
  /** e.g. Supreme Court, Parliament, SEBI, High Court */
  category: string;
  title: string;
  summary: string;
  /** e.g. Landmark Judgment, New Law, Amendment */
  tag: string;
  /** Hex color for the tag badge, e.g. #DC2626 */
  tagColor: string;
  published?: boolean;
}