# lib/api-zod/src/generated/types/newsArticle.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NewsArticle {
  id: number;
  date: string;
  category: string;
  title: string;
  summary: string;
  tag: string;
  tagColor: string;
  published: boolean;
  createdAt: string;
}